edited project (portfolio)

+++ demo/src/main/java/ir/tejarattrd/oms/demo/demo/Entity/Order.java
@@ -1,8 +1,8 @@
 package ir.tejarattrd.oms.demo.demo.Entity;
-import ir.tejarattrd.oms.demo.demo.Entity   .Symbol;
+
 import jakarta.persistence.*;
+import java.math.BigDecimal; // **ایمپورت جدید**
 import java.time.LocalDateTime;
-/*-------------------------------------------------------------------------------- */
 
 @Entity
 @Table(name = "orders")
@@ -21,47 +21,42 @@ public class Order {
     private Symbol symbol;
 
     @Column(name = "order_type")
-    private String orderType; // مثلا "buy" یا "sell"
+    private String orderType;
 
+    // **تغییر کلیدی: نوع داده به BigDecimal تغییر کرد**
     @Column(name = "quantity")
-    private Double quantity;
+    private BigDecimal quantity;
 
-    @Column(name = "total_price", precision = 18)
-    private Double totalPrice;
+    // **تغییر کلیدی: نوع داده به BigDecimal تغییر کرد**
+    @Column(name = "total_price")
+    private BigDecimal totalPrice;
 
     @Column(name = "order_date")
     private LocalDateTime orderDate;
 
     @Column(name = "status")
-    private String status; // مثلا "pending", "completed", "cancelled"
+    private String status;
 
     @PrePersist
     protected void onCreate() {
         orderDate = LocalDateTime.now();
     }
 
-    // Getters and Setters
+    // Getters and Setters (با نوع BigDecimal)
     public Long getOrderId() { return orderId; }
     public void setOrderId(Long orderId) { this.orderId = orderId; }
-
     public Customer getCustomer() { return customer; }
     public void setCustomer(Customer customer) { this.customer = customer; }
-
     public Symbol getSymbol() { return symbol; }
     public void setSymbol(Symbol symbol) { this.symbol = symbol; }
-
     public String getOrderType() { return orderType; }
     public void setOrderType(String orderType) { this.orderType = orderType; }
-
-    public Double getQuantity() { return quantity; }
-    public void setQuantity(Double quantity) { this.quantity = quantity; }
-
-    public Double getTotalPrice() { return totalPrice; }
-    public void setTotalPrice(Double totalPrice) { this.totalPrice = totalPrice; }
-
+    public BigDecimal getQuantity() { return quantity; }
+    public void setQuantity(BigDecimal quantity) { this.quantity = quantity; }
+    public BigDecimal getTotalPrice() { return totalPrice; }
+    public void setTotalPrice(BigDecimal totalPrice) { this.totalPrice = totalPrice; }
     public LocalDateTime getOrderDate() { return orderDate; }
     public void setOrderDate(LocalDateTime orderDate) { this.orderDate = orderDate; }
-
     public String getStatus() { return status; }
     public void setStatus(String status) { this.status = status; }
 }
